fix comment, fill password field in rdp with decrypted password
When doing RDP without 2step enabled, the username and password fields are shown aswell. Previously, the password field was empty eventhough the CP had the password stored internally. Now the password field is filled with the decrypted password - correct number of *'s - to indicate that the password is already present.